requirements.txt
torch
torchvision

torchmetrics
pillow
scikit-learn
mypy
ruff
numpy
onnxscript
